Complete the temperature conversion function

+++ Clima/WeatherDataModel.swift
@@ -11,7 +11,7 @@ import UIKit
 class WeatherDataModel {
 
     //Declare your model variables here
-    var temperature : Int = 0
+    var temperature : Double = 0
     var condition : Int = 0
     var city : String = ""
     var weatherIconName : String = ""
